[KAIZEN-0] flyttet saga-test-utils til egen fil
